fix: ให้ avg-resolution ของ asset reliability ปัดเป็นชั่วโมงแบบเดียวกับทั้งระบบ
จุดสุดท้ายที่ยังปัดสองรอบ: query asset reliability ใช้ ROUND(AVG(min),0)
ปัด avg นาทีเป็นจำนวนเต็มก่อนหารเป็นชั่วโมง (68.5→69 → 1.15 → 1.2) ขณะที่
การ์ด summary กับ MTTR ช่างได้ 1.1 สำหรับงานชุดเดียวกัน. ตอนนี้ทั้ง 3 metric
ใช้ pipeline เดียวกันเป๊ะ: ROUND(AVG(min),1) → cast float → round(x/60,1)
ครั้งเดียว.

- ReportRepository: asset avg_resolution_minutes เป็น ROUND(AVG(min),1)
- ReportService::mapAssetReliabilityRow: cast avgMinutes เป็น float
- test: เพิ่ม regression asset 17+120 นาที ต้องได้ 1.1 (36 tests pass)

ผล: AST-AC-0001 บนหน้า/PDF/Excel เปลี่ยน 1.2 → 1.1 ตรงกับการ์ด+ช่าง

+++ app/Services/ReportService.php
@@ -204,7 +204,9 @@ class ReportService
 
     private function mapAssetReliabilityRow(array $row): array
     {
-        $avgMinutes = (int) ($row['avg_resolution_minutes'] ?? 0);
+        // float (not int): same single-rounding pipeline as summary + technician MTTR — keep the
+        // .5-minute precision so avg-resolution rounds to hours consistently across the whole report.
+        $avgMinutes = (float) ($row['avg_resolution_minutes'] ?? 0);
         $status = (string) ($row['status'] ?? 'active');
 
         return [
